Graphics.GLUtil.GLError

Description

Miscellaneous utilities for dealing with OpenGL errors.

Documentation

printError :: IO ()

Check OpenGL error flags and print them on stderr.

printErrorMsg :: String -> IO ()

Check OpenGL error flags and print them on stderr with the given message as a prefix. If there are no errors, nothing is printed.

throwError :: IO ()

Throw an exception if there is an OpenGL error.

data GLError

An exception type for OpenGL errors.

throwErrorMsg :: String -> IO ()

Throw an exception if there is an OpenGL error. The exception's error message is prefixed with the supplied String.