Graphics.GLUtil.JuicyTextures

Description

Uses the JuicyPixels package to load images that are then used to create OpenGL textuers.

Documentation

readTexInfo :: FilePath -> (forall a. IsPixelData a => TexInfo a -> IO b) -> IO (Either String b)

Load a TexInfo value from an image file, and supply it to a user-provided function. Supported image formats include png, jpeg, bmp, and gif. See readTexture for most uses.

readTexture :: FilePath -> IO (Either String TextureObject)

Load a TextureObject from an image file. Supported formats include png, jpeg, bmp, and gif.