Graphics.GLUtil.Linear

Description

Support for writing Linear types to uniform locations in shader programs.

Documentation

class AsUniform t where

A type class for things we can write to uniform locations in shader programs. We can provide instances of this class for types from Linear without introducing orphan instances.

Minimal complete definition

Nothing

Methods

asUniform :: t -> UniformLocation -> IO ()

asUniform :: Uniform t => t -> UniformLocation -> IO ()