Graphics.GLUtil.ShaderProgram

Description

Convenience interface for working with GLSL shader programs. Provides an interface for setting attributes and uniforms.

The ShaderProgram type

data ShaderProgram

Representation of a GLSL shader program that has been compiled and linked.

Constructors

ShaderProgram

Fields

attribs :: Map String (AttribLocation, VariableType)

uniforms :: Map String (UniformLocation, VariableType)

program :: Program

Simple shader programs utilizing a vertex shader and a fragment shader

simpleShaderProgram :: FilePath -> FilePath -> IO ShaderProgram

Load a ShaderProgram from a vertex shader source file and a fragment shader source file. The active attributes and uniforms in the linked program are recorded in the ShaderProgram.

simpleShaderProgramWith :: FilePath -> FilePath -> (Program -> IO ()) -> IO ShaderProgram

Load a ShaderProgram from a vertex shader source file and a fragment shader source file. The active attributes and uniforms in the linked program are recorded in the ShaderProgram. The supplied IO function is applied to the new program after shader objects are attached to the program, but before linking. This supports the use of bindFragDataLocation to map fragment shader outputs.

simpleShaderExplicit :: FilePath -> FilePath -> ([String], [String]) -> IO ShaderProgram

Load a ShaderProgram from a vertex and fragment shader source files. the third argument is a tuple of the attribute names and uniform names that will be set in this program. If all attributes and uniforms are desired, consider using loadShaderProgram.

simpleShaderProgramBS :: ByteString -> ByteString -> IO ShaderProgram

Load a ShaderProgram from vertex and fragment shader source strings. The active attributes and uniforms in the linked program are recorded in the ShaderProgram.

simpleShaderProgramWithBS :: ByteString -> ByteString -> (Program -> IO ()) -> IO ShaderProgram

Load a ShaderProgram from vertex and fragment shader source strings. See simpleShaderProgramWith for more information.

simpleShaderExplicitBS :: ByteString -> ByteString -> ([String], [String]) -> IO ShaderProgram

loadShaderFamily :: FilePath -> IO ShaderProgram

Load a shader program from vertex, geometry, and fragment shaders that all share the same root file name and the various conventional extensions: ".vert", ".geom", and ".frag". If a specific file doesn't exist, such as a geometry shader, it is skipped. For instance, loadShaderFamily "simple" will load and compile, "simple.vert" and "simple.frag" if those files exist.

Explicit shader loading

loadShaderProgram :: [(ShaderType, FilePath)] -> IO ShaderProgram

Load a ShaderProgram from a list of individual shader program files. The active attributes and uniforms in the linked program are recorded in the ShaderProgram

loadShaderProgramWith :: [(ShaderType, FilePath)] -> (Program -> IO ()) -> IO ShaderProgram

loadShaderProgramBS :: [(ShaderType, ByteString)] -> IO ShaderProgram

Load a ShaderProgram from a list of individual shader program source strings. The active attributes and uniforms in the linked program are recorded in the ShaderProgram

loadShaderProgramWithBS :: [(ShaderType, ByteString)] -> (Program -> IO ()) -> IO ShaderProgram

Working with ShaderProgram parameters

getAttrib :: ShaderProgram -> String -> AttribLocation

Get the AttribLocation associated with a named vertex attribute.

enableAttrib :: ShaderProgram -> String -> IO ()

Enable a named vertex attribute.

setAttrib :: ShaderProgram -> String -> IntegerHandling -> VertexArrayDescriptor a -> IO ()

Set a named vertex attribute's IntegerHandling and VertexArrayDescriptor.

setUniform :: AsUniform a => ShaderProgram -> String -> a -> IO ()

Set a named uniform parameter associated with a particular shader program.

getUniform :: ShaderProgram -> String -> UniformLocation

Get the UniformLocation associated with a named uniform parameter.