Graphics.GLUtil.Viewport

Description

Helpers for working with OpenGL viewports.

Documentation

withViewport :: MonadIO m => Position -> Size -> m a -> m a

withViewport pos sz m runs the action m after setting the viewport with the given Position and Size. The viewport is reset to its original state after the action is run, and the result of the action is returned.